file-system/requirements.txt
mcp[cli]>=1.7.1
PyPDF2>=3.0.1
python-docx>=1.1.2
openpyxl>=3.1.5
python-pptx>=1.0.2
pillow>=11.2.1
pymupdf>=1.25.5
pdfplumber>=0.11.6
tabula-py>=2.10.0
ebooklib>=0.19
beautifulsoup4>=4.13.4
striprtf>=0.0.29
fastmcp>=2.2.6 